+++ src/main/java/duke/Parser.java
@@ -1,3 +1,7 @@
+package duke;
+
+import duke.command.*;
+
 public class Parser {
 
     public static Command parse(String command) throws DukeException {
