+++ src/main/java/duke/command/AddCommand.java
@@ -1,5 +1,12 @@
+package duke.command;
+
+import duke.DukeException;
+import duke.Storage;
+import duke.Ui;
+import duke.task.TaskList;
+
 public class AddCommand extends Command {
-    private String input;
+    private final String input;
 
     public AddCommand(String input) {
         this.input = input;
