+++ src/main/java/duke/command/DeleteCommand.java
@@ -1,7 +1,14 @@
-public class DoneCommand extends Command {
+package duke.command;
+
+import duke.DukeException;
+import duke.Storage;
+import duke.Ui;
+import duke.task.TaskList;
+
+public class DeleteCommand extends Command {
     private final int taskNo;
 
-    public DoneCommand(String taskNo) throws DukeException {
+    public DeleteCommand(String taskNo) throws DukeException {
         try {
             this.taskNo = Integer.parseInt(taskNo);
         } catch (NumberFormatException e) {
@@ -10,7 +17,7 @@ public class DoneCommand extends Command {
     }
 
     public void execute(TaskList tasks, Ui ui, Storage storage) {
-        tasks.taskDone(taskNo);
+        tasks.deleteTask(taskNo);
     }
 
     public boolean isExit() {
