+++ src/main/java/duke/command/UnknownCommand.java
@@ -1,3 +1,10 @@
+package duke.command;
+
+import duke.DukeException;
+import duke.Storage;
+import duke.Ui;
+import duke.task.TaskList;
+
 public class UnknownCommand extends Command {
     public void execute(TaskList tasks, Ui ui, Storage storage) throws DukeException {
         throw new DukeException("OOPS!!! I'm sorry, but I don't know what that means :-( " +
