+++ src/main/java/duke/task/Task.java
@@ -1,3 +1,5 @@
+package duke.task;
+
 public class Task {
     public String taskName;
     public boolean isDone;
